requirements.txt
#requirements.txt for pip installation: <pip install -r requirements.txt>
Django
django-taggit